Migrate tests/ directory to automake.
Add Makefile.am to build test binary and use it for the check target.

The #include'd source files test_*.c must be in EXTRA_DIST because
they cannot be compiled using test_SOURCES, but must be distributed
with the package.  This should be fixed sometime to build separate
object files for each and link them together for the test binary.