Fix reading past EOF on windows.
On windows, a file is implicitly opened in text mode unless the
O_BINARY flag is passed to open().  While reading from the file, CRLF
line endings are then automatically converted to just LF newlines.
This causes the number of bytes that can be read from the file to
differ from the file length reported by fstat().

As a consequence, the ihex_rs_from_mem() loop for counting the records
will read from a buffer that contains less bytes than expected
(although properly sized), caused by the missing CR characters.
Subsequently it can lead to a wrong number of records reported in the
rs->ihrs_count field, if the uninitialized garbage in the buffer
happens to contain additional IHEX_CHR_RECORDMARK bytes.

Open file in O_BINARY mode if supported (useful on Windows).